+++ src/main/kotlin/com/bitmoxie/monorepochangedprojects/MonorepoChangedProjectsPlugin.kt
@@ -1,4 +1,4 @@
-package com.douglan.monorepochangedprojects
+package com.bitmoxie.monorepochangedprojects
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project